refactor: wrapper style 파일 위치 조정
