+++ src/App.jsx
@@ -1,7 +1,7 @@
 import React from "react";
 
 import Header from "@/components/header/Header";
-import ContentWrapper from "@/components/Wrapper.styled";
+import ContentWrapper from "@/components/wrapper/Wrapper.styled";
 import ProductList from "@/pages/home/components/product-list/ProductList";
 
 function App() {
