+++ src/components/header/Header.jsx
@@ -1,7 +1,7 @@
 import { css } from "@emotion/react";
 
 import StyledHeader from "@/components/header/Header.styled";
-import HeaderWrapper from "@/components/Wrapper.styled";
+import HeaderWrapper from "@/components/wrapper/Wrapper.styled";
 import LogoLink from "@/components/header/logo-link/LogoLink";
 import Menu from "@/components/header/menu/Menu";
